Rename DeclarationTextBox to DeclarationFormInteractor
.. and move from Reko.Gui.Windows.Controls to Reko.Gui.Windows.Forms
