+++ src/Gui/Gui.csproj
@@ -289,7 +289,7 @@
     </Compile>
     <Compile Include="Windows\Controls\TextViewLayout.cs" />
     <Compile Include="Windows\Controls\TextViewPainter.cs" />
-    <Compile Include="Windows\Controls\DeclarationTextBox.cs" />
+    <Compile Include="Windows\Forms\DeclarationFormInteractor.cs" />
     <Compile Include="Windows\Forms\DeclarationForm.cs">
       <SubType>Form</SubType>
     </Compile>
